Start moving a few of the fundamental system services into their own processes and files
The eventual goal is for task() to become a hlt loop, but for now this is not possible because the kernel
never explicitly preempts out of itself, so task0 must explicitly yield. It would - of course - be possible
to do something like

L: hlt
   yield
   jmp L

as a workaround.

For now, split up these core system processes:
 - scheduler
 - awaker
 - collector
 - deleter

And let task0 be just

L: yield