Teach the build system how to produce a bootable hard disk image
This adds a few requirements on the build environment (namely a docker container running in privileged mode, or a bare Linux install with loop mounting support).
Is it - however - not required, as the "os.iso" image is still produced.
Eventually, the goal is to phase the ISO out, and just produce a bootable HD image with a sane filesystem layout.