build: intelligently move files to the main filesystem vs. initrd
/initrd is a relatively scarce resource (64 files max). This is arguably a self-imposed
limitation, but it is already quite larger than it need be. This was done so that one could
insert user binaries in /initrd before having a proper hard disk boot strategy. Now that we have one,
move most binaries off /initrd and into <mainfs>/apps

A few select binaries remain in /initrd, namely
  - init, because - well - duh
  - mount, in order to be allowed to proceed with mounting the main filesystem
  - ls, because it's useful to look around