syscalls: enable a system call to be implemented directly via _helper
The default is for the system call manager to call _helper, which is a stub routine
that goes from SyscallRequest to the individual arguments to the system call and dispatches
those to _handler, the real implementation.

This is generally OK, but it prevents the system call from changing the value of EIP/EFLAGS
or any other register/flag that the kernel may allow to modify in the future. So, allow the system
call to skip _handler and implement itself directly via _helper, which exposes the raw
SyscallRequest.