test: add a "libcheckup" testing harness and a simple test case
libcheckup is the testing harness for puppy. It's a small hand-rolled testing infrastructure
with minimal overhead and (hopefully) extremely simple to use.

Improvements will happen over time (e.g. helpers to instantiate new tests, run a harness made of multiple tests)
and it's likely that new kernel/userspace APIs will be needed to write interesting tests

For now, write a simple tests that checks that one can malloc memory and safely write to it