vmm: introduce COW pages
Introduce the notion of COW pages, where a page is mapped read-only, but with a magic "COW" bit flagged, such that
when a page fault occurs while trying to write to the page, a copy is made and the copy is then writable.

This is usable to implement some kind of "process cloning", when a somewhat exact replica of a process' memory is made,
and the new copy is executed as a separate process (not yet implemented)

In order to support this, enable WP, i.e. prevent even the OS from writing to read-only pages. This required a few tweaks
to the VMM layer.