boot: implement a simple kernel command-line parser
This parses options of the form <NAME>=<VALUE> and then tries to match "name" to a known list and in turn
to a value. Currently this is very manual as there is only one option (loglevel = 0 for no logging, 255 for full logging).

Once we have a few more options, this should probably be automated by means of a configuration file.

Add a grub boot entry to load the kernel with no logging, using this new mechanism