synch: add a test case to reproduce a bug with process killing
If a process is killed while it's waiting on a mutex (or a semaphore for that matter...), the kernel
moves the TSS information in order to execute the termination, but it leaves the process suspended so that
it can be resumed when the synchronization object becomes available. This is done so that the object's queue does not
end up containing stale process objects.

However, this means that a process could end up forever stuck waiting to collect a process that is waiting on a resource
held, e.g. by a terminated process. While the process would never really get scheduled again, it would consume resources
(e.g. memory, files, ...) that could not be freed without a reboot.

This is hardly optimal - but the fix is not trivial in the current design. So, add a test case.