synch: add a WaitQueue helper type
The goal is for semaphores, messages and any other kind of wait (maybe except sleep..) to happen on a WaitQueue instead
of building new process states per wait type. Instead, any entity that wants to implement "process waiting" would just use
a WaitQueue and call wait and wake primitives as it needs to.

This commit is NFC as nothing adopts the new type