memory: add an API to allocate a physical memory range
allocContiguousPages() will try to get a "range" of physical pages allocated
Since this operation happens at purely the physical page level, it does not assign
virtual addresses, nor give any process ownership of the memory. As such it is the
responsibility of the caller to access and manage the returned pages as needed.