vfs: s/Volume/IDEVolume/g
Rename the "Volume" class to "IDEVolume" because that's what it really is: a Volume off an IDE disk

There's room to improve even more here, e.g. a class for a volume off a physical disk with support for disk partitions and scanning
that would abstract DiskScanner off of IDE, but that's left as an exercise for the reader.