newlib: proof of concept newlib port to Puppy
This commit begins the process of porting newlib to Puppy. The newlib build process is a tad bit byzantine, so this commit
drops prebuilt newlib .a files in order to expedite the build and not overly complicate the build machinery (*).

This is without excessive loss of value as the actual system call implementation for Puppy is then provided
in a separate interface.cpp file that is compiled into an extra .a file to be linked into applications.

A minimal set of system calls is (not fully) ported - sufficient to implement a program that prints a string and exits. This is
then used to (very much ad-hoc) build and install a newlibdemo program in the mainfs.

For the above reasons, this is mostly a proof of concept and work to follow is largely on providing proper implementations
of the APIs that newlib requires + possibly making kernel changes where needed to support a sane C execution environment.

(*) For the interested readers, following the directions at https://wiki.osdev.org/Porting_Newlib mostly works, provided one uses
newlib tip of tree (8f1259a6ef03edfa0f8e8b28fcaa13410b4e6b10) + automake-1.11.6 + autoconf-2.64 and patches the automake script
as described at https://github.com/raspberrypi/noobs/issues/470 (but the line number will be different) and symlinks aclocal-1.11 to aclocal
in the automake/autoconf install path