newlib: initial prototype implementation of relative paths
This hooks up newlib fopen() with support for relative paths. The resolution algorithm goes kinda like this:

if the path begins with / and does not contain any . characters (TODO: check for dots), then it's an absolute path - leave alone
if either of this is not satisfied, prefix with the cwd (which better be an absolute path - TODO: check for that)
Take the combined path and split into '/' separated components, discard all components that are just '.' (i.e. current directory);
this leaves us, with a list of paths - navigate through them
if any of them is '..' and it's not the first or second of the list, then mark the previous path component as "unused" (TODO: nested dot dot)
merge all path components that are not marked "unused" with a / between them and an initial /, return as the "absolute path"

This works for many cases of interest, e.g.

/tmp/./file.txt
/system/../tmp/file.txt
/tmp/../tmp/file.txt
/tmp/../../../tmp/file.txt

But it would not work for cases like

/system/config/../../tmp/file.txt
/tmp//file.txt

This algorithm is also not hooked up to any other calls that take file paths in newlib's interface (e.g. stat, opendir, ...)

Part of https://github.com/egranata/puppy/issues/23