newlib: improve relative path management
This improves the prototypical algorithm to support multiple '..' separators and to do a more reliable job at detecting absolute paths

While I can't personally come up with a test case that would fail the current algorithm, it is not the "correct" solution. A better model would be
to treat the sequence of path components as a queue language, where each component means "push myself", except "." means no-op and ".." means pop the
previous element off - and then just concatenate all the leftover components with '/' separators - and ending up with an empty queue is an error

In theory this model would make it simple to even implement the very old school "..." (go up two levels) shortcut (not that it's a good idea...)

Next steps - given that now there are some test cases - are:

in newlib
- implement the queue model;
- extend this to other filesystem-y operations;

in kernel
- cwd should be an absolute path - validate that;
- cwd should also exist - would be nice to validate that.

and - last - should CWD be in kernel at all? The only argument I can think of to support that is that it is probably generally useful and being in kernel
makes it easy for a child process to inherit the parent's value

More work towards https://github.com/egranata/puppy/issues/23