elf: prototype the ability to load ELF dynamic libraries
This adds a (temporary, so in the 200 range) system call that can load an ELF dynamic library based on load-time relocation, perform a few
simple relocations - and at least do well enough to be able to call a simple function in that dynamic library.

The code itself is a tad bit hacky as it tries to coexist with the code to load an ELF executable with no relocation support (and also
it's a WIP with only a few basic relocation types support) and the test app isn't much more production ready.

In order to clean this up:
- move to userspace (e.g. load a /system/libs/libELFLoader.so at a well-known absolute location);
- provide dlsym() as an entry point, do not rely on hardcoding a function pointer (*)
- better error handling and sanity checking;

(*) GCC does not even always link at that address, so YMMV...

One (huge) step towards https://github.com/egranata/puppy/issues/37