time: move to a TimeManager API to support scheduling and clocks - and move away from the PIT
This moves a bunch of time-management logic away from the PIT driver onto a new TimeManager API
TimeManager is now responsible for collecting on-time-IRQ tasks, executing them, and keeping an accurate uptime count for the system

As part of this cleanup, starting using the local APIC timer as the source of timing events instead of the PIT
Part of this change is moving the "ms/tick" value to 5 from its previous count of 10.
Setting this value too low (e.g. 1Hz) puts too much strain on the system, but 5Hz is both a reasonable choice and worth doing as a test