apps: prototype text editor app
This is an early prototype of an ed-like text editor for Puppy.

Thus far it can only append lines to a new file and save to disk. Future goals would be adding the ability to edit/replace individual lines and load existing text files.

In general, it would be nice to have a properly usable text editor, but to do so would require some kind of curses support for full-screen apps.