build: experimental: use GCC specs to setup a default option set for Puppy userspace builds
This is an experiment with using specs file in order to obtain something fairly close to a i686-puppy-gcc without having to customize the GCC source tree.
When run against a built and mounted Puppy image (at $PWD/out/mnt), this essentially allows to say gcc.sh -c file.cpp -o file.o and invoke a compiler that is custom tailored to Puppy.

There are probably other ways to this same goal that do not involve the dark arts, but this mechanism allows for a fair bit of customization of the compiler environment and it qualifies
as a good experiment in what's possible. This commit also introduces a -keep-mounted option to build.py, which is meant to be used to keep the OS disk image mounted at the end of the build
process (this would be useful if one were, e.g. compiling extra programs and needed the Puppy libraries and includes - or just looking around the freshly baked OS image).