exceptions: implement GPF recovery
This is a - rather fragile and low-level - mechanism to support controlled recovery from a general protection fault

What happens is that - via pushGPFRecovery() - one can ask the kernel to not panic on a in-kernel GPF, but instead return to the provided address
The assumption is that the returned-to instruction will know how to safely escape the fault and return to a proper state. This is - as said - somewhat dangerous and fragile as:
- it relies on detailed knowledge of the caller's worldview;
- it loses state information as to the cause of the fault;
- it is racy with, e.g. faults happening in IRQs.

Quite a bit of sanity checking is provided in the implementation in order to try and offset some of these issues as possible.

One big TODO item is left open: if the recovery does not really recover, and the same (or a different) GPF happens before the recovery address is popped, the kernel will end up in an endless loop.
The solution would be to only allow a recovery address to be used N (==1 probably being the (only?) sane value) before going straight to a kernel panic