build: stop using loop devices during the build
There is at least one interesting build environment where one can't actually use loopback devices.
In order to support such, stop manually mounting the disk image file and instead assemble it from parts.

This requires a bit of hackery:
- have a prebuilt boot sector ready (*);
- copy GRUB modules and support files by hand;
- assemble the final disk image from a FAT volume + a header portion

(*) currently, this is overwriting the partition table - which is not really desirable;
one solution would be to just to copy the boot code leaving the partition table intact.
This will only start mattering if we want to increase the image file size over the current 64MB value