process: allow a process to be rescheduled on its way out of a system call
The kernel itself is non-preemtible - such that a process that's running a system call may receive a timer IRQ, but not be "kicked out" by the scheduler
because it is in a non-preemptible kernel region (we figure this out by the address the process is at, anything >= 3GB is kernel land).

Currently, this means that a process running repeated system calls may - given some sheer luck - end up not preempted for a relatively long amount of time, if
all the timer IRQs happen to catch it during a system call. This is all the more possible because we do not batch disk I/O at the moment, meaning disk reads and writes happen in real time.

Improve this situation by introducing a "ready for reschedule" flag. The idea is that - should the timer IRQ decide that a process is ripe for kicking out, but cannot be because it is running in kernel space,
it will mark it with this new flag. At the very end of a system call, the kernel will check this flag and - should it be true - will invoke the scheduler to pick a new running process

Because this only happens on the way out of a system call, the rest of the kernel can still safely assume it will not be interrupted