vmm: fix an issue with tracking of physical pages
If the kernel freed a page backed by actual storage, and then reassigned the same page, but zeromapped, and freed this latter mapping,
this would cause a panic because the "frompmm" flag wasn't being cleared. as a result, the kernel would try and free a physical page for
the virtual mapping.