awaker: use a WaitQueue to control the awaker process
Only keep the awaker task running if a process is in need of its services;
This is suboptimal as a process may ask to sleep for an hour, and this
would keep the awaker up and running for that long, whereas the awaker itself
could sleep for almost an hour.. but who would wake the awaker then?

A good solution to this would involve moving the wake-up logic
in the scheduler - same as the exiting logic.

OTOH, this + lowering the scheduling priority of the awaker does bring some
improvement.