process: argv should be an array, not a string
This modifies the kernel and userspace APIs to execute a process from taking (const char*, const char*)
into taking (const char*, char**) with the expectation that argv[0] equals program_path.

This makes the interface more naturally UNIX-y.

For now this comes with a bunch of hacks all around:
- proctable does not copy process arguments anymore;
- libShellSupport allocates a bunch of memory that is never freed;
- shell builtins are hacked to work, but the design could be better

It does not solve, but definitely is the major step towards https://github.com/egranata/puppy/issues/57