mm: add fail-safety to allocating physical memory pages
I would like to add a cache in the PhysicalPageManager such that the manager has a list of
a few free pages that can quickly be given out without a list scan. This has the potential to
break things in the physical page allocator, which is A Bad Thing.

Harden the rest of the kernel (just the VirtualPageManager really) against this occurrence by making
sure that either a valid page was obtained, or there is a kernel panic