synch: remove legacy semaphore/mutex APIs
This gets rid of the legacy APIs based around semget() and mutexget().
This allows getting rid of the system calls, as well as the entire infrastructure around lifetime management
(which is now handled by the VFS layer).
It also allows removing the Handletables - and the space saved is used to make a larger file handle table instead

This deletes >800 lines of code which is an amazing win
Also, it fixes https://github.com/egranata/puppy/issues/69