i386: add support for x86 MTRR
Using MTRR allows configuring regions of physical memory with several properties about how the hardware manages accessing that memory

The original goal was to setup write-combining for the framebuffer memory. Benchmarking has shown this to be not desirable on at least one system of interest.
So, keep the support code in i386/ but do not actually allow the framebuffer to write-combine.

This might be good to add as a boot option at some point.