synch: Introduce wait tokens
Each process owns a "wait token" (aka, a uint64_t), which starts at 0 at process spawn
When a process goes to sleep or enters a WaitQueue, the current value of the token is stashed away.

When a process is awoken, the token's value is incremented. But the process is only awoken by a
sleep timeout or wait queue wake if the process's current token value matches what was stored at
the beginning.

This is foundational infrastructure to allow, e.g. a process to both sleep and wait on a queue,
or to wait on multiple queues.