build: add -keep-user flag
This flag builds a new kernel (and core libs) but keeps userspace intact.
-rebuild-app is ignored in this mode

The idea is that if the kernel is the only thing that is changed, and the kernel <--> userspace interface
is not changed (e.g. no new system calls, no file format tweaks), the same userspace will run on the previous
kernel