syscalls: add syscall action bits
System calls return a 32-bit value such that:
- if the LSB is 0, this is to be interpreted as "success" and the value is a result
(e.g. a pointer to a new memory region, or a count - shifted by 1 if need be);
- otherwise, the result is an error, and - shifted by 1 - the rest of the value is the
error code

This adds the notion of using 4 more bits of the "error" state as "action bits" for the
system call machinery itself to decode and act upon.

A system call can return such an "action code" as error + setting one or more of the bits to
a known action code, and the syscall_irq handler will do something about that depending
on the bits.

The action bits are not meant for userland consumption - and the syscall itself might
end up being a success or a fail depending on the specific bits used