kernel: setup version/signature info
The OS signature is the branch + commit hash this OS image is being built from
The OS version is the short hash of the commit being treated as a number

The signature is meant to be unique and universally identify a given build of Puppy (minus local deltas)
such that anyone could check out that exact commit and obtain the same OS bits.
The version is a shorthand for the above - with no guarantees - and not even any promise of
monotonicity, but can be used to match a number to an OS build if needed.

The idea is to use the signature to reproduce a build, features to discover properties of the OS
from userspace and the version as a shorthand for display purposes