log: setup a log callback and prepare to make use of it
The idea behind this is to post all kernel messages past a certain point in very early boot to a message queue.
Userspace could then pick these messages up and handle them as it sees fit. In practice, this would go
through a "logd" system which would receive kernel messages as well as messages from userland processes, and
mix them up in a global system-wide log. It's likely fine to do this skipping very early boot. Or maybe not and
I will have to revisit this part of the design somehow.

The only real nuance here is to avoid recursion in the logging API.